:zap: :: complete qrCodeScreen

+++ app/src/main/java/com/staker4wapper/flick_kiosk/presentation/Screen/home/Product.kt
@@ -2,6 +2,6 @@ package com.staker4wapper.flick_kiosk.presentation.Screen.home
 
 data class Product(
     val image: Int? = null,
-    val price: Int,
+    val price: Long,
     val name: String
 )
